Added 1 puzzle to Road to Foxborough
I added Bill Belichick as NPC to the Tunnel. I have the dialogue I want
in the game in the code but it’s not working. I added an “an every turn
rule” in that makes so you can only talk to Bill while holding the
correct items but that’s not working too well

"Road To Foxborough" by Avery Serna

Outsideroom is a room. Printed name is "Outside"

North of outsideroom is Hallway.

East of Hallway is EastHallway.
EastHallway is a room. Printed name is "East Hallway"

West of Hallway is WestHallway.
WestHallway is a room. Printed name is "West Hallway"

North of EastHallway is SoutheastLR.
SoutheastLR is a room. Printed name is "Southeast Locker Room"

North of WestHallway is SouthwestLR.
SouthwestLR is a room. Printed name is "Southwest Locker Room"

North of SoutheastLR is NortheastLR.
NortheastLR is a room. Printed name is "Northeast Locker Room"

North of SouthwestLR is NorthwestLR.
NorthwestLR is a room. Printed name is "Northeast Locker Room"

East of SoutheastLR is Equipmentroom1.
Equipmentroom1 is a room. Printed name is "Equipment Room 1"

East of NortheastLR is Equipmentroom2.
Equipmentroom2 is a room. Printed name is "Equipment Room 2"

West of NorthwestLR is QBroom.
QBroom is a room. Printed name is "QB Room"

West of SouthwestLR is Conferenceroom.
Conferenceroom is a room. Printed name is "Conference Room"

North of NorthwestLR is Tunnel.
Tunnel is a room.
Bill Belichick is a man in Tunnel. "Coach Bill is standing in the
darkness of the tunnel, wearing the same, cut Patriots hoodie, giving
inspiring words to the players running by him."
Understand "talk to Bill Belichick" as talking.

An every turn rule:
	if the player is wearing helmet:
		if the player is wearing jersey:
			if the player is wearing shoulder pads:
				if the player is wearing pants:
					if the player is wearing shoes:
						if the player is wearing gloves;
						talk to Bill Belichick

"Hey kid. I need a big favor of you. We really need this win today so I
need you to sneak into the Falcon's locker room and steal their
playbook. You can use the secret underground passage. The entrance is
in the QB room, use the passcode 'psi 12.5' to gain entrance. Don't let
us down".

North of NortheastLR is Tunnel.

North of Tunnel is GilletteStadium.
GilletteStadium is a room. Printed name is "Gillette Stadium"

North of GilletteStadium is Falcontunnel.
Falcontunnel is a room. Printed name is "Falcon's Tunnel"

North of Falcontunnel is FalconLR.
FalconLR is room. Printed name is "Falcon Locker Room"